arkanBot/requirements.txt
python-telegram-bot==20.6
dateparser==1.0.0
python-dotenv==0.19.0
datetime